renamed folder and added namespace to scripts
