+++ Parry/Assets/Scripts/LevelGeneration/Segment.cs.meta
