--- Parry/Assets/Scripts/SegmentS/KillZone.cs
@@ -1,26 +1,0 @@
-using UnityEngine;
-
-public class KillZone : MonoBehaviour
-{
-    [SerializeField] private SegmentSpawner segmentSpawner;
-    
-    private void OnTriggerEnter(Collider other)
-    {
-        if (!HasSegmentTag(other)) return;
-
-        Segment segmentComponent = GetSegmentComponent(other);
-        
-        segmentComponent.Reset();
-        segmentSpawner.SpawnNext();
-    }
-
-    private bool HasSegmentTag(Collider other)
-    {
-        return other.CompareTag("Segment");
-    }
-
-    private Segment GetSegmentComponent(Component other)
-    {
-        return other.GetComponent<Segment>();
-    }
-}
